Inline single-use helpers into their sole call sites
HasCount only had one conformer (ChartSeries), and utcDateFormatter
only had one consumer (UTCTimestampText). Move both out of the
Utilities folder into the files that actually use them.
