--- Sources/Scout/UI/Utilities/HasCount.swift
@@ -1,17 +1,0 @@
-//
-// Copyright 2025 Mikhail Kasianov
-//
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file or at
-// https://opensource.org/licenses/MIT.
-
-protocol HasCount {
-    associatedtype Count: AdditiveArithmetic
-    var count: Count { get }
-}
-
-extension Collection where Element: HasCount {
-    var total: Element.Count {
-        reduce(.zero) { $0 + $1.count }
-    }
-}
