--- Sources/Scout/UI/Utilities/UTCDateFormatter.swift
@@ -1,16 +1,0 @@
-//
-// Copyright 2026 Mikhail Kasianov
-//
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file or at
-// https://opensource.org/licenses/MIT.
-
-import Foundation
-
-/// A shared UTC date formatter used across detail views.
-let utcDateFormatter: DateFormatter = {
-    let formatter = DateFormatter()
-    formatter.timeZone = TimeZone(secondsFromGMT: 0)
-    formatter.dateFormat = "dd.MM.y, HH:mm"
-    return formatter
-}()
